+++ frontend/src/components/index.ts
@@ -49,7 +49,7 @@ export * from 'components/ChallengeInfoWithUser';
 export * from 'components/Record';
 export * from 'components/ErrorFallbackMain';
 export * from 'components/Header/ErrorFallbackLoginButton';
-export * from 'components/ErrorFallbackSubscriptionButton';
-export * from 'components/ErrorFallbackNotificationMessage';
+export * from 'components/Header/ErrorFallbackSubscriptionButton';
+export * from 'components/Header/ErrorFallbackNotificationMessage';
 export * from 'components/MembersPopover';
 export * from 'components/CheckSuccessCycle';
